moved lib-stuff to `lib`
