+++ lib/_BaseModel.js
@@ -17,7 +17,10 @@
  * Copyright (c) OWASP Foundation. All Rights Reserved.
  */
 
-class CycloneDXObject {
+/**
+ * @internal
+ */
+class _BaseModel {
   validateType (name, value, expectedType, required = false) {
     if (!value) {
       if (required) { throw new ReferenceError(name + ' is required') }
@@ -59,4 +62,4 @@ class CycloneDXObject {
   }
 }
 
-module.exports = CycloneDXObject
+module.exports = _BaseModel
