+++ lib/models/AttachmentText.js
@@ -17,7 +17,7 @@
  * Copyright (c) OWASP Foundation. All Rights Reserved.
  */
 
-const CycloneDXObject = require('./CycloneDXObject')
+const CycloneDXObject = require('../_BaseModel')
 
 class AttachmentText extends CycloneDXObject {
   constructor (contentType, text, encoding) {
