+++ lib/models/Bom.js
@@ -20,7 +20,7 @@
 const builder = require('xmlbuilder')
 const uuid = require('uuid')
 const Component = require('./Component')
-const CycloneDXObject = require('./CycloneDXObject')
+const CycloneDXObject = require('../_BaseModel')
 const Metadata = require('./Metadata')
 const Tool = require('./Tool')
 const program = require('../../package.json')
