+++ lib/models/Component.js
@@ -19,7 +19,7 @@
 
 const parsePackageJsonName = require('parse-packagejson-name')
 const { PackageURL } = require('packageurl-js')
-const CycloneDXObject = require('./CycloneDXObject')
+const CycloneDXObject = require('../_BaseModel')
 const LicenseChoice = require('./LicenseChoice')
 const HashList = require('./HashList')
 const ExternalReferenceList = require('./ExternalReferenceList')
