+++ lib/models/ExternalReference.js
@@ -17,7 +17,7 @@
  * Copyright (c) OWASP Foundation. All Rights Reserved.
  */
 
-const CycloneDXObject = require('./CycloneDXObject')
+const CycloneDXObject = require('../_BaseModel')
 
 class ExternalReference extends CycloneDXObject {
   constructor (type, url, comment) {
