+++ lib/models/HashList.js
