+++ lib/models/License.js
@@ -18,7 +18,7 @@
  */
 
 const AttachmentText = require('./AttachmentText')
-const CycloneDXObject = require('./CycloneDXObject')
+const CycloneDXObject = require('../_BaseModel')
 
 class License extends CycloneDXObject {
   get id () {
