+++ lib/models/LicenseChoice.js
@@ -21,7 +21,7 @@ const spdxLicenses = require('../../res/spdx-licenses.json')
 const fs = require('fs')
 const License = require('./License')
 const AttachmentText = require('./AttachmentText')
-const CycloneDXObject = require('./CycloneDXObject')
+const CycloneDXObject = require('../_BaseModel')
 
 class LicenseChoice extends CycloneDXObject {
   constructor (pkg, includeLicenseText = true) {
