+++ lib/models/Metadata.js
@@ -18,7 +18,7 @@
  */
 
 const Component = require('./Component')
-const CycloneDXObject = require('./CycloneDXObject')
+const CycloneDXObject = require('../_BaseModel')
 const OrganizationalEntity = require('./OrganizationalEntity')
 
 class Metadata extends CycloneDXObject {
