+++ lib/models/OrganizationalContact.js
@@ -17,7 +17,7 @@
  * Copyright (c) OWASP Foundation. All Rights Reserved.
  */
 
-const CycloneDXObject = require('./CycloneDXObject')
+const CycloneDXObject = require('../_BaseModel')
 
 class OrganizationalContact extends CycloneDXObject {
   constructor (name, email, phone, objectName) {
