+++ lib/models/OrganizationalEntity.js
@@ -17,7 +17,7 @@
  * Copyright (c) OWASP Foundation. All Rights Reserved.
  */
 
-const CycloneDXObject = require('./CycloneDXObject')
+const CycloneDXObject = require('../_BaseModel')
 const OrganizationalContact = require('./OrganizationalContact')
 
 class OrganizationalEntity extends CycloneDXObject {
