+++ lib/models/Tool.js
@@ -17,7 +17,7 @@
  * Copyright (c) OWASP Foundation. All Rights Reserved.
  */
 
-const CycloneDXObject = require('./CycloneDXObject')
+const CycloneDXObject = require('../_BaseModel')
 
 class Tool extends CycloneDXObject {
   constructor (vendor, name, version, hashes = []) {
