+++ tests/lib/models/Metadata.test.js
@@ -19,12 +19,12 @@
  * Copyright (c) OWASP Foundation. All Rights Reserved.
  */
 
-const Bom = require('../../src/models/Bom')
-const Component = require('../../src/models/Component')
-const Metadata = require('../../src/models/Metadata')
-const OrganizationalContact = require('../../src/models/OrganizationalContact')
-const OrganizationalEntity = require('../../src/models/OrganizationalEntity')
-const Tool = require('../../src/models/Tool')
+const Bom = require('../../../lib/models/Bom')
+const Component = require('../../../lib/models/Component')
+const Metadata = require('../../../lib/models/Metadata')
+const OrganizationalContact = require('../../../lib/models/OrganizationalContact')
+const OrganizationalEntity = require('../../../lib/models/OrganizationalEntity')
+const Tool = require('../../../lib/models/Tool')
 
 test('Model: Metadata / Format: XML', () => {
   const result = createMetadata().toXML()
